+++ test/basic/expr_attr.sv
@@ -1,0 +1,4 @@
+module top;
+    wire a, b, c, d, e;
+    assign a = b ? (* ternary *) ~^ (* unary *) c : d & (* binary *) e;
+endmodule